# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WalletWatch API",
    "description": "API documentation for WalletWatch, a Budgeting and Expense Tracking Application made with Express.js and TypeScript.",
    "contact": {
      "email": "seifusama@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4650",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/expenses": {
      "get": {
        "tags": ["Expenses"],
        "summary": "Get all expenses for the logged-in user",
        "description": "Returns a list of all expenses for the logged-in user",
        "operationId": "getExpenses",
        "responses": {
          "200": { "description": "List of expenses retrieved successfully" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal Server error" }
        }
      },
      "post": {
        "tags": ["Expenses"],
        "summary": "Create a new expense",
        "description": "Creates a new expense for the logged-in user",
        "operationId": "createExpense",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Expense"
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Expense created successfully" },
          "400": { "description": "Invalid input" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal Server error" }
        }
      }
    },
...
    "/budgets": {
      "get": {
        "tags": ["Budgets"],
        "summary": "Get all budgets for the logged-in user",
        "description": "Returns a list of all budgets for the logged-in user",
        "operationId": "getBudgets",
        "responses": {
          "200": { "description": "List of budgets retrieved successfully" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal Server error" }
        }
      },
      "post": {
        "tags": ["Budgets"],
        "summary": "Create a new budget",
        "description": "Creates a new budget for the logged-in user",
        "operationId": "createBudget",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Budget"
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Budget created successfully" },
          "400": { "description": "Invalid input" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal Server error" }
        }
      }
    },
    "/goals": {
      "get": {
        "tags": ["Goals"],
        "summary": "Get all goals for the logged-in user",
        "description": "Returns a list of all goals for the logged-in user",
        "operationId": "getGoals",
        "responses": {
          "200": { "description": "List of goals retrieved successfully" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal Server error" }
        }
      },
      "post": {
        "tags": ["Goals"],
        "summary": "Create a new goal",
        "description": "Creates a new goal for the logged-in user",
        "operationId": "createGoal",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Goal"
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Goal created successfully" },
          "400": { "description": "Invalid input" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal Server error" }
        }
      }
    },
    "/reports": {
      "get": {
        "tags": ["Reports"],
        "summary": "Get reports for the logged-in user",
        "description": "Returns a list of reports for the logged-in user",
        "operationId": "getReports",
        "responses": {
          "200": { "description": "List of reports retrieved successfully" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal Server error" }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Expense": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "64a7f0c2e4b0c8b1a1e4d2f3"
          },
          "userId": { "type": "string", "example": "64a7f0c2e4b0c8b1a1e4d2f3" },
          "amount": { "type": "number", "example": 50.75 },
          "category": { "type": "string", "example": "Food" },
          "description": { "type": "string", "example": "Grocery shopping" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Budget": {
        "type": "object",
        "properties": {
          "_id": { "type": "string", "example": "64a7f0c2e4b0c8b1a1e4d2f3" },
          "userId": { "type": "string", "example": "64a7f0c2e4b0c8b1a1e4d2f3" },
          "type": {
            "type": "string",
            "example": "Monthly",
            "enum": ["monthly", "weekly"]
          },
          "amount": { "type": "number", "example": 500 },
          "startDate": { "type": "string", "format": "date-time" },
          "endDate": { "type": "string", "format": "date-time" },
          "spent": { "type": "number", "example": 150 },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "Goal": {
        "type": "object",
        "properties": {
          "_id": { "type": "string", "example": "64a7f0c2e4b0c8b1a1e4d2f3" },
          "userId": { "type": "string", "example": "64a7f0c2e4b0c8b1a1e4d2f3" },
          "title": { "type": "string", "example": "New Laptop" },
          "targetAmount": { "type": "number", "example": 1500 },
          "currentAmount": { "type": "number", "example": 300 },
          "deadline": { "type": "string", "format": "date-time" },
          "status": {
            "type": "string",
            "enum": ["active", "achieved", "failed"]
          },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}